pyproject.toml
[project]
name = "computer-use-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic",
    "anyio",
    "bandit",
    "bs4",
    "colorama",
    "ftfy",
    "icecream",
    "langchain-ai",
    "langchain-community",
    "mypy",
    "networkx",
    "openai",
    "pandas",
    "playwright",
    "pyautogui",
    "pydeps",
    "pysnooper",
    "python-dotenv",
    "ruff",
    "tenacity",
    "types-beautifulsoup4",
    "types-pyautogui",
    "unimport",
]
